requirements.txt
appdirs==1.4.3
Django==1.11.1
django-crispy-forms==1.6.1
django-filter==1.0.2
django-taggit==0.22.1
djangorestframework==3.6.3
Markdown==2.6.8
packaging==16.8
pkg-resources==0.0.0
pytz==2017.2
six==1.10.0
whitenoise==3.3.0
